pyproject.toml
# ============================================================================
# Python Project Configuration for APPSIERRA (Poetry)
# ============================================================================
# This file contains project metadata and tool configurations for Poetry.
# Poetry Documentation: https://python-poetry.org/docs/pyproject/
# ============================================================================

[tool.poetry]
name = "appsierra"
version = "0.1.0"
description = "APPSIERRA - Sierra Chart Trading Monitor with PyQt6"
authors = ["APPSIERRA Team"]
readme = "README.md"
license = "Proprietary"
keywords = ["trading", "sierra-chart", "dtc", "pyqt6", "monitoring"]

# Package information
packages = [
    { include = "config" },
    { include = "core" },
    { include = "panels" },
    { include = "services" },
    { include = "ui" },
    { include = "utils" },
    { include = "widgets" },
]

# ============================================================================
# Dependencies
# ============================================================================

[tool.poetry.dependencies]
python = ">=3.11,<3.15"
PyQt6 = "^6.0.0"
pyqtgraph = "^0.13.0"
pydantic = "^2.0.0"
sqlmodel = "^0.0.14"
structlog = "^24.0.0"
blinker = "^1.7.0"
orjson = "^3.9.0"
colorio = "^0.12.18"
colorspacious = "^1.1.2"
psutil = "^7.0.0"
numpy = "^2.0.0"
pyyaml = "^6.0.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.0"
pre-commit = "^3.0.0"
pytest = "^7.0.0"
pytest-cov = "^4.0.0"
pytest-qt = "^4.0.0"
mypy = "^1.0.0"
alembic = "^1.13.0"
watchdog = "^3.0.0"
pyinstaller = "^6.0.0"
memory-profiler = "^0.61.0"
py-spy = "^0.3.14"

# ============================================================================
# Build System
# ============================================================================
pytest-xdist = "^3.8.0"
pycln = "^2.6.0"
vulture = "^2.14"
radon = "^6.0.1"
safety = "^3.7.0"
coverage = "^7.11.3"
hypothesis = "^6.147.0"
tox = "^4.32.0"
types-pyyaml = "^6.0.12.20250915"
types-requests = "^2.32.4.20250913"
cx-freeze = "^8.4.1"
loguru = "^0.7.3"
yappi = "^1.7.3"
snakeviz = "^2.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# ============================================================================
# Ruff Configuration
# ============================================================================
# Note: Main ruff config is in ruff.toml
# This section provides minimal overrides if needed

[tool.ruff]
# Use ruff.toml as the primary configuration file
extend = "ruff.toml"

# ============================================================================
# Pytest Configuration
# ============================================================================

[tool.pytest.ini_options]
# Test discovery patterns
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

# Output options
addopts = [
    "-v",                    # Verbose output
    "--strict-markers",      # Strict marker checking
    "--tb=short",           # Shorter traceback format
    "--disable-warnings",   # Disable warnings
    "-ra",                  # Show summary of all test outcomes
]

# Markers for organizing tests
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "ui: marks tests that require Qt/GUI",
]

# Minimum Python version
minversion = "7.0"

# ============================================================================
# Coverage Configuration
# ============================================================================

[tool.coverage.run]
# Files to include in coverage
source = [
    "config",
    "core",
    "panels",
    "services",
    "ui",
    "utils",
    "widgets",
]

# Files to omit from coverage
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/.venv/*",
    "*/venv/*",
    "setup.py",
]

# Branch coverage
branch = true

# Parallel mode
parallel = true

[tool.coverage.report]
# Minimum coverage percentage
fail_under = 70

# Show missing lines
show_missing = true

# Skip covered files
skip_covered = false

# Skip empty files
skip_empty = true

# Precision for coverage percentage
precision = 2

# Exclude lines from coverage
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.coverage.html]
directory = "htmlcov"

# ============================================================================
# Bandit Security Configuration
# ============================================================================

[tool.bandit]
# Directories to exclude from security scanning
exclude_dirs = [
    ".venv",
    "venv",
    "tests",
    ".git",
    "__pycache__",
]

# Tests to skip (if needed)
skips = [
    # "B101",  # assert_used (assert is ok in tests)
]

# Severity level for reporting
# Options: all, low, medium, high
level = "medium"

# Confidence level for reporting
# Options: all, low, medium, high
confidence = "medium"

# ============================================================================
# MyPy Type Checking Configuration (optional)
# ============================================================================

[tool.mypy]
# Python version
python_version = "3.11"

# Strictness
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Gradual typing for PyQt6 app
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
disallow_any_generics = false  # PyQt6 uses many generics
strict_optional = true
no_implicit_optional = true

# Error output
show_error_context = true
show_column_numbers = true
show_error_codes = true
pretty = true
color_output = true

# Import discovery
namespace_packages = true
explicit_package_bases = true
ignore_missing_imports = false  # Explicit per-module overrides below

# Warnings
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

# Cache
cache_dir = ".mypy_cache"
incremental = true

# Ignore missing imports for these libraries
[[tool.mypy.overrides]]
module = [
    "PyQt6.*",
    "pyqtgraph.*",
]
ignore_missing_imports = true

# ============================================================================
# Black Formatter Configuration (alternative to ruff format)
# ============================================================================
# Note: Using ruff format instead, but keeping this for reference

[tool.black]
line-length = 120
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.git
  | \.venv
  | venv
  | __pycache__
  | \.pytest_cache
  | \.mypy_cache
  | build
  | dist
)/
'''

# ============================================================================
# isort Configuration (alternative to ruff's isort)
# ============================================================================
# Note: Using ruff's isort instead, but keeping this for reference

[tool.isort]
profile = "black"
line_length = 120
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
split_on_trailing_comma = true
skip_gitignore = true
known_third_party = ["PyQt6", "pyqtgraph"]
known_first_party = ["config", "core", "panels", "services", "ui", "utils", "widgets"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
